+++ src/modules/addLike.js
@@ -20,11 +20,9 @@ export const getBtn = async () => {
   const likeBtn = document.querySelectorAll('.like');
   likeBtn.forEach((btn) => {
     btn.addEventListener('click', (e) => {
-      const id = e.target.id.split('Like')[0]; // This line is to return the actual ID removing the appended 'Like' at the end
+      const id = e.target.id.split('Like')[0];
 
       likeFn(id);
-      // window.location.reload();
-      // setTimeout(window.location.reload.bind(location), 300);
       setTimeout(() => window.location.reload(), 300);
     });
   });
